refactor hook useFavorite
