+++ src/components/Header/Project.jsx
@@ -2,32 +2,33 @@ import React, { useState, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import { useFetchCharacter } from '../../hooks/useFetchCharacter'
 import PropTypes from 'prop-types'
+import { useFavorites } from '../../hooks/useFavorites'
 
 function Project() {
   const { id } = useParams()
   const { personInfo, isLoading } = useFetchCharacter(id)
   const { name, status, species, origin, location, gender } = personInfo
   const [imageLoading, setImageLoading] = useState(true)
-  const [isFavorite, setIsFavorite] = useState(false)
+  const { favoriteCards } = useFavorites()
 
   useEffect(() => {
     const storedFavorites = JSON.parse(localStorage.getItem('favorites')) || []
     const isAlreadyFavorite = storedFavorites.some((card) => card.id === id)
     setIsFavorite(isAlreadyFavorite)
-  }, [id])
+  }, [id, favoriteCards])
+
+  const [isFavorite, setIsFavorite] = useState(false)
 
   const handleImageLoaded = () => {
     setImageLoading(false)
   }
 
-  const handleAddToFavorites = () => {
-    setIsFavorite(!isFavorite)
-
+  const handleAddOrRemoveFromFavorites = () => {
     const storedFavorites = JSON.parse(localStorage.getItem('favorites')) || []
-
     if (isFavorite) {
       const updatedFavorites = storedFavorites.filter((card) => card.id !== id)
       localStorage.setItem('favorites', JSON.stringify(updatedFavorites))
+      setIsFavorite(false)
     } else {
       const newFavorite = {
         id,
@@ -43,6 +44,7 @@ function Project() {
         'favorites',
         JSON.stringify([...storedFavorites, newFavorite])
       )
+      setIsFavorite(true)
     }
   }
 
@@ -71,7 +73,7 @@ function Project() {
         <p>Location: {location?.name}</p>
         <p>Gender: {gender}</p>
         <button
-          onClick={handleAddToFavorites}
+          onClick={handleAddOrRemoveFromFavorites}
           className="btn-favorites"
           style={{
             backgroundColor: isFavorite ? '#7950f2' : '#6350d3',
